refactor: move async-mutex into service/tools/
Its only consumer is cadence-export.ts, so colocate it there.
